switches tab header logo to correct mock
